src/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.1
pycparser==2.20
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
XlsxWriter==1.3.7
